requirements.txt
alembic==1.9.2
anyio==3.6.2
click==8.1.3
fastapi==0.89.1
fastapi-pagination==0.11.3
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.5
pydantic==1.10.4
python-dotenv==0.21.1
sniffio==1.3.0
SQLAlchemy==2.0.1
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
